cros_ui_test: Do not overwrite a legitimate resolv.conf
Test to see if resolv.conf is already pointng to our local server
and that a backup file already exists.  This would indicate that
perhaps the last test was interrupted (crash) and that we do not
want to overwrite our backup with this bogus copy.

BUG=chromium-os:18823
TEST=Visual only.  Watching tree while one builder continues to exhibit
this issue.

Reviewed-on: https://gerrit.chromium.org/gerrit/21802
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>

+++ client/cros/cros_ui_test.py
@@ -130,8 +130,19 @@ class UITest(pyauto_test.PyAutoTest):
             resolv_test = os.path.join(resolv_dir, self._resolv_test_file)
             resolv_bak = os.path.join(resolv_dir, self._resolv_bak_file)
             resolv_contents = 'nameserver 127.0.0.1'
-            # Back up the current resolv.conf.
-            os.rename(resolv, resolv_bak)
+            # Test to make sure the current resolv.conf isn't already our
+            # specially modified version.  If this is the case, we have
+            # probably been interrupted while in the middle of this test
+            # in a previous run.  The last thing we want to do at this point
+            # is to overwrite a legitimate backup.
+            if (utils.read_one_line(resolv) == resolv_contents and
+                os.path.exists(resolv_bak)):
+                logging.error('Current resolv.conf is setup for our local '
+                              'server, and a backup already exists!  '
+                              'Skipping the backup step.')
+            else:
+              # Back up the current resolv.conf.
+              os.rename(resolv, resolv_bak)
             # To stop flimflam from editing resolv.conf while we're working
             # with it, we want to make the directory -r-x-r-x-r-x.  Open an
             # fd to the file first, so that we'll retain the ability to
